changed int

+++ src/Main.java
@@ -5,7 +5,7 @@ public class Main {
         int amount = 2473;
         System.out.println("Сумма пополнения:" + amount);
         if (amount > 1000) {
-            int replenishment = (int) (balance + amount + amount / 100);
+            int replenishment = (balance + amount + amount / 100);
             System.out.println("Вы пополнили счет на " + amount + " рублей - бонус равен " + amount / 100 + " рубля/ей, итоговая сумма на балансе " + replenishment + " рубля/ей");
         }
         else  {
